+++ inst/sql/sql_server/analyses/1331.sql
@@ -1,6 +1,6 @@
 -- 1331	Proportion of people with at least one visit_detail record outside a valid observation period
 --
--- stratum_1:   Proportion to 6 decimal places
+-- stratum_1:   Proportion
 -- stratum_2:   Number of people with a record outside a valid observation period (numerator)
 -- stratum_3:   Number of people in visit_detail (denominator)
 -- count_value: Flag (0 or 1) indicating whether any such records exist
@@ -30,7 +30,7 @@ FROM
 SELECT 
 	1331 AS analysis_id,
 	CASE WHEN vdt.person_count != 0 THEN 
-		CAST(CAST(1.0*op.person_count/vdt.person_count AS NUMERIC(7,6)) AS VARCHAR(255)) 
+		CAST(CAST(1.0*op.person_count/vdt.person_count AS FLOAT) AS VARCHAR(255)) 
 	ELSE 
 		CAST(NULL AS VARCHAR(255))
 	END AS stratum_1,
