+++ inst/sql/sql_server/analyses/1332.sql
@@ -1,6 +1,6 @@
 -- 1332	Proportion of visit_detail records outside a valid observation period
 --
--- stratum_1:   Proportion to 6 decimals places
+-- stratum_1:   Proportion
 -- stratum_2:   Number of visit_detail records outside a valid observation period (numerator)
 -- stratum_3:   Number of visit_detail records (denominator)
 -- count_value: Flag (0 or 1) indicating whether any such records exist
@@ -30,7 +30,7 @@ FROM
 SELECT 
 	1332 AS analysis_id,
 	CASE WHEN vdt.record_count != 0 THEN 
-		CAST(CAST(1.0*op.record_count/vdt.record_count AS NUMERIC(7,6)) AS VARCHAR(255)) 
+		CAST(CAST(1.0*op.record_count/vdt.record_count AS FLOAT) AS VARCHAR(255)) 
 	ELSE 
 		CAST(NULL AS VARCHAR(255)) 
 	END AS stratum_1,
